requirements.txt
flask==1.1.1
bs4==0.0.1
flask-material==0.1.1
gensim-sum-ext==0.1.2
nltk==3.4.5
spacy==2.2.3
sumy==0.8.1
python-dotenv==0.12.0
readtime==1.1.1
gunicorn==20.0.4; python_version > '3.0'
gunicorn==19.10.0; python_version < '3.0'
https://github.com/explosion/spacy-models/releases/download/en_core_web_sm-2.2.5/en_core_web_sm-2.2.5.tar.gz#en_core_web_sm

